done with creating and reading files

+++ controllers/admin.js
@@ -1,6 +1,6 @@
 const Product = require('../models/product');
 const { validationResult } = require("express-validator");
-const mongoose = require('mongoose');
+const deleteFile = require('../util/deleteImage');
 
 exports.getAddProduct = (req, res, next) => {
   res.render('admin/edit-product', {
@@ -88,7 +88,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const updatedImage = req.file;
   const updatedDesc = req.body.description;
 
   const errors = validationResult(req)
@@ -100,7 +100,6 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDesc,
         _id: prodId
@@ -118,7 +117,10 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
+      if(updatedImage){
+        deleteFile(product.imageUrl)
+        product.imageUrl = updatedImage.path;
+      }
       return product.save()
         .then(result => {
           res.redirect('/admin/products');
@@ -151,13 +153,19 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
-    .then(() => {
-      res.redirect('/admin/products');
-    })
-    .catch(err => {
-      const error = new Error(err)
-      error.httpStatusCode = 500
-      return next(error)
-    });
+  Product.findById(prodId).then(product=>{
+    if(!product){
+      return next(new Error("Product not found"))
+    }
+
+    deleteFile(product.imageUrl)
+    return Product.deleteOne({ _id: prodId, userId: req.user._id })
+  })
+  .then(() => {
+    res.redirect('/admin/products');
+  }).catch(err => {
+    const error = new Error(err)
+    error.httpStatusCode = 500
+    return next(error)
+  });
 };
